@@ -8,33 +8,27 @@
     "build": "ng build --prod",
     "test": "ng test",
     "lint": "ng lint",
-    "postinstall": "ng build -prod --aot",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
     "@agm/core": "^1.0.0-beta.2",
-    "@angular/cdk": "^5.0.4",
-    "@angular/cli": "^1.6.3",
-    "@angular/common": "^5.1.3",
-    "@angular/compiler": "^5.1.3",
-    "@angular/compiler-cli": "^5.1.3",
-    "@angular/core": "^5.1.3",
-    "@angular/forms": "^5.1.3",
-    "@angular/http": "^5.1.3",
-    "@angular/material": "^5.0.4",
-    "@angular/platform-browser": "^5.1.3",
-    "@angular/platform-browser-dynamic": "^5.1.3",
-    "@angular/router": "^5.1.3",
-    "bootstrap": "^4.0.0-beta.3",
+    "@angular/common": "^5.2.1",
+    "@angular/core": "^5.2.1",
+    "@angular/forms": "^5.2.1",
+    "@angular/http": "^5.2.1",
+    "@angular/platform-browser": "^5.2.1",
+    "@angular/platform-browser-dynamic": "^5.2.1",
+    "@angular/router": "^5.2.1",
     "core-js": "^2.5.3",
+    "ng4-loading-spinner": "^1.1.1",
     "ngmap": "^1.18.4",
-    "ngx-cookie-service": "^1.0.10",
-    "npm-template": "^0.1.1",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.19"
+    "ngx-cookie-service": "^1.0.10"
   },
   "devDependencies": {
+    "@angular/cli": "^1.6.5",
+    "@angular/compiler": "^5.2.1",
+    "@angular/compiler-cli": "^5.2.1",
     "@angular/language-service": "^5.1.3",
     "@types/jasmine": "^2.8.3",
     "@types/jasminewd2": "^2.0.3",
@@ -51,7 +45,8 @@
     "protractor": "^5.2.2",
     "ts-node": "^4.1.0",
     "tslint": "^5.8.0",
-    "typescript": "^2.4.2"
+    "typescript": "^2.4.2",
+    "zone.js": "^0.8.20"
   },
   "engines": {
     "node": "6.11.0",
