@@ -2,40 +2,23 @@ import { Injectable } from '@angular/core';
 import { Http, Response, Headers, RequestOptions } from '@angular/http';
 import 'rxjs/add/operator/toPromise';
 import { Subject, Observable } from 'rxjs/Rx'
+import 'rxjs/add/operator/map'
 
 
 @Injectable()
 export class WebService {
-//  BASE_URL = 'http://api.markl.fi';
-  BASE_URL = 'https://polar-river-59836.herokuapp.com';
+  BASE_URL = 'http://api.markl.fi';
 
   private result = [];
-
-
-  private resultAll = new Subject();
-  private resultCity = new Subject();
-
-  observations = this.resultAll.asObservable();
-  citydata = this.resultCity.asObservable();
 
   constructor(private http: Http) { }
 
   getObservations() {
-    this.http.get(this.BASE_URL).subscribe(res => {
-      this.result = res.json();
-      this.resultAll.next(this.result[0]);
-    }, error => {
-      console.error(error);
-    });
+    return this.http.get(this.BASE_URL).map(res => res.json()[0]);
   };
 
   getCityData(city) {
-    this.http.get(this.BASE_URL + "/" + city).subscribe(res => {
-      this.result = res.json();
-      this.resultCity.next(this.result);
-    }, error => {
-      console.error(error);
-    });
+    return this.http.get(this.BASE_URL + "/" + city).map(res => res.json()[0]);
   };
 
   async postObservation(postData) {
